+++ src/ContentProcessor.ts
@@ -1,9 +1,9 @@
 ﻿import k8s = require('@kubernetes/client-node');;
 import { isNullOrUndefined } from "util";
-import { AddedTypes } from "./AddedTypes";
 import { logger } from "./LoggerWrapper";
 import { DeployReplica, IRootObject } from "./RequestDefinition";
-
+import { TemplateValidator } from './TemplateValidator';
+import { DiffCalculator } from './DiffCalculator'
 export class ContentProcessor {
 
     public static async TryUpdateConfig(message: string): Promise<string> {
@@ -102,103 +102,5 @@ export class ContentProcessor {
             logger.info(`failed to get extra data error ${JSON.stringify(error)}`);
             throw (error);
         });
-    }
-}
-
-export class TemplateValidator {
-    public static ValidateContent(content: IRootObject) {
-        let returnValue = true;
-        logger.info(`validating content ${JSON.stringify(content)}`);
-
-        if (isNullOrUndefined(content)) {
-            logger.error("null content");
-            returnValue = false;
-        }
-        else if (isNullOrUndefined(content.request)
-            || isNullOrUndefined(content.request.operation)
-            || (content.request.operation !== "CREATE"
-                && content.request.operation !== "UPDATE")) {
-
-            logger.error("invalid incoming operation");
-            returnValue = false;
-        }
-        else if (isNullOrUndefined(content.kind)
-            || (content.kind !== "AdmissionReview" && content.kind !== "Testing")) {
-
-            logger.error("invalid incoming kind");
-            returnValue = false;
-        }
-        else if (isNullOrUndefined(content.request.object)
-            || isNullOrUndefined(content.request.object.spec)) {
-
-            logger.error("missing spec in template");
-            returnValue = false;
-        }
-
-        logger.info(`succesfully validated content ${JSON.stringify(content)}`);
-        return returnValue;
-    }
-}
-
-export class DiffCalculator {
-    public static CalculateDiff(content: IRootObject, extraData: DeployReplica): object {
-
-        if (isNullOrUndefined(content)) {
-            logger.error("null content");
-            return null;
-        }
-
-        /* tslint:disable */
-        logger.info(`calculating diff`);
-        const updatedContent: IRootObject = JSON.parse(JSON.stringify(content));
-
-        let updateTarget: object;
-
-        try {
-            updateTarget = updatedContent.request.object.spec.template.spec;
-            logger.info(`updating request.object.spec.template.spec`);
-        }
-        catch (ex) {
-            updateTarget = updatedContent.request.object.spec;
-            logger.info(`updating request.object.spec`);
-        }
-
-        if (updateTarget["initContainers"]) {
-            Array.prototype.push.apply(updateTarget["initContainers"], AddedTypes.init_containers());
-        } else {
-            updateTarget["initContainers"] = AddedTypes.init_containers();
-        }
-
-        if (updateTarget["volumes"]) {
-            Array.prototype.push.apply(updateTarget["volumes"], AddedTypes.volumes());
-        } else {
-            updateTarget["volumes"] = AddedTypes.volumes();
-        }
-
-        const length = updateTarget["containers"].length;
-
-        for (let i = 0; i < length; i++) {
-            if (updateTarget["containers"][i].env) {
-                Array.prototype.push.apply(updateTarget["containers"][i].env, AddedTypes.env(extraData));
-            } else {
-                updateTarget["containers"][i].env = AddedTypes.env(extraData);
-            }
-
-            if (updateTarget["containers"][i].volumeMounts) {
-                Array.prototype.push.apply(updateTarget["containers"][i].volumeMounts, AddedTypes.volume_mounts());
-            } else {
-                updateTarget["containers"][i].volumeMounts = AddedTypes.volume_mounts();
-            }
-        }
-
-        const jsonDiff = [
-            {
-                op: "replace",
-                path: "/spec",
-                value: updatedContent.request.object.spec
-            }];
-        /* tslint:enable */
-        logger.info(`determined diff ${JSON.stringify(jsonDiff)}`);
-        return jsonDiff;
     }
 }
